yaft: add DEAD_LINES hack for framebuffers that can't draw the full extent
DevTerm is missing 2-4px along the top of the screen due to a kernel
driver bug. This workaround tells Yaft not to use them. It's a WIP (in
particular, it only supports lines along the left edge of the FB, which
is what gets drawn as the top in DT), and it's currently disabled
because I don't have a font that conveniently divides into the reduced
screen size.