yaft: support ITU format for 8/24-bit colour
This is the format that uses : rather than ; to separate subparams. It
also adds support for direct colour with or without a colourspace ID,
although if it's not unambiguous (the ; format is used and there are
more than five trailing subparams) it assumes a lack of colourspace ID,
as this is generally safer.

The old (legacy xterm) format is still supported, but the terminfo and
termcap files now recommend ITU format.