yaft: support modified arrow keys
We can't get the VT to send us different sequences for these, so
instead, if any modifiers are being held down, we scan our input for the
relevant escape codes and, if found, emit the modified ones instead.

The modified codes are added to terminfo and chosen to match xterm's
behaviour.