yaft: test blit performance on startup and adjust drawing strategy accordingly
In non-rotated mode, the framebuffer and drawing surface are in
alignment, and each line can be drawn with CELL_HEIGHT calls to memcpy()
spanning the whole line. In rotated mode, each *cell* has to be drawn as
CELL_WIDTH individual memcpy()s, which is usually much slower. Copying
out the whole screen using line-by-line memcpy() can be faster if large
amounts of the screen are dirtied, but how much faster is hard to
predict in advance.

With this change, on startup, yaft briefly tests the performance of
different drawing strategies, and uses that to figure out how much of
the screen needs to be dirtied for it to be worthwhile redrawing the
whole thing. On the R01 this results in a 2-3x speedup for the common
case where large amounts of the screen have been dirtied.