yaft: add basic support for DECSET/DECRST 1049
This is the "alternate drawing buffer" or "ti/te" mode, a nonscrolling
display that temporarily replaces the main scrolling display without
forgetting its contents. It's typically used by fullscreen interfaces
like less, atuin, and whatnot.

This is a fairly basic implementation that probably doesn't do nearly as
much as it should and has some visible graphical glitches, but it is at
least usable.